+++ backend-rust/CHANGELOG.md
@@ -4,7 +4,27 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-Database schema version: 20
+### Added
+
+- `baker_id` to `AccountAddress` in transaction events.
+
+### Fixed
+
+- Fixed bug when setting `--log-level` causing some project logs to be filtered out, and filter dependencies to `info` level. This behavior can be overwritten using `RUST_LOG`.
+
+## [0.1.40] - 2025-03-26
+
+Database schema version: 22
+
+### Fixed
+
+- Rename `ValidatorScoreParametersChainUpdatePayload` to `ValidatorScoreParametersUpdate` for backwards compatibility on dotnet
+- Fix the `AccountByAddress::transactionCount` query to return the number of transactions the account has been involved in or
+  affected by. Expose the `AccountByAddress::nonce` query to return the account nonce.
+
+## [0.1.39] - 2025-03-25
+
+Database schema version: 22
 
 ### Added
 
@@ -13,8 +33,11 @@ Database schema version: 20
 
 ### Fixed
 
+- Add database migration 22 fixing some passive delegators that had no `delegated_restake_earnings` value set in the database.
 - Change `Versions::backend_versions` to `Versions::backend_version`.
 - When `effective_time` is zero then it translates into `Block::slot_time`.
+- Split up migration 18 into several SQL transactions to avoid timeouts for long running migrations.
+- Change CLI option `--log-level` to only apply for logs produced directly from this project, instead of including every dependency.
 
 ## [0.1.38] - 2025-03-21